Revert "Combine runtimes"
