+++ cumulus/parachain-template/node/src/rpc.rs
@@ -7,7 +7,7 @@
 
 use std::sync::Arc;
 
-use subspace_runtime::execution::{opaque::Block, AccountId, Balance, Index as Nonce};
+use parachain_template_runtime::{opaque::Block, AccountId, Balance, Index as Nonce};
 
 use sc_client_api::AuxStore;
 pub use sc_rpc::{DenyUnsafe, SubscriptionTaskExecutor};
